+++ src/main/java/com/stemm/pubsub/service/post/entity/Post.java
@@ -1,11 +1,12 @@
-package com.stemm.pubsub.post.entity;
+package com.stemm.pubsub.service.post.entity;
 
 import com.stemm.pubsub.common.BaseEntity;
-import com.stemm.pubsub.user.entity.User;
+import com.stemm.pubsub.service.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import static jakarta.persistence.EnumType.STRING;
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
@@ -13,16 +14,12 @@ import static lombok.AccessLevel.PROTECTED;
 @NoArgsConstructor(access = PROTECTED)
 @Getter
 @Entity
-@Table(indexes = @Index(name = "idx_comment_created_date", columnList = "created_date"))
-public class Comment extends BaseEntity {
+@Table(indexes = @Index(name = "idx_post_created_date", columnList = "created_date"))
+public class Post extends BaseEntity {
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    @Column(name = "comment_id")
+    @Column(name = "post_id")
     private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "user_id")
@@ -32,13 +29,16 @@ public class Comment extends BaseEntity {
     @Column(nullable = false)
     private String content;
 
-    @Column(nullable = false)
-    private Long likeCount;
+    private String imageUrl;
 
-    public Comment(Post post, User user, String content, Long likeCount) {
-        this.post = post;
+    @Enumerated(STRING)
+    @Column(nullable = false)
+    private Visibility visibility;
+
+    public Post(User user, String content, String imageUrl, Visibility visibility) {
         this.user = user;
         this.content = content;
-        this.likeCount = likeCount;
+        this.imageUrl = imageUrl;
+        this.visibility = visibility;
     }
 }
