+++ src/main/java/com/stemm/pubsub/service/post/entity/PostLike.java
@@ -1,7 +1,7 @@
-package com.stemm.pubsub.post.entity;
+package com.stemm.pubsub.service.post.entity;
 
 import com.stemm.pubsub.common.BaseEntity;
-import com.stemm.pubsub.user.entity.User;
+import com.stemm.pubsub.service.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
